-Ajout de la nouvelle RivLibLibrary pour Linux, l'ancienne ne prend pas en compte les tirs sans écho -Visualisation: 	filtrage des voxels par intervalle, étirement des couleurs sur intervalle 	ajout du mode orthographique à la caméra

+++ trunk/VoxeLidarGUIFX/src/main/java/fr/ird/voxelidar/Configuration.java
@@ -566,7 +566,7 @@ public class Configuration {
                                     String inequality = e.getAttributeValue("inequality");
                                     String value = e.getAttributeValue("value");
 
-                                    filters.add(new Filter(variable, Double.valueOf(value), Filter.getConditionFromString(inequality)));
+                                    filters.add(new Filter(variable, Float.valueOf(value), Filter.getConditionFromString(inequality)));
                                 }
                             }
                         }
